Initial update to POI integration plugin
Currently, most of the old features should work properly, only there is
now a limitation that it will only work with objective waypoints that
are in your current zone, in particular with the ones on the watch
frame. I will likely change it so that the ones on the world map work
properly, but this will be in a separate commit.