Add a function to fetch coordinates safely
For the coordinate feed and coordinate block. this function can be used
instead of the heavier player position function to get the current
player's coordinates. This means that depending on the current map zoom,
the 'coordinates' will display the position on that map, which is what
we'd expect.

The arrow and rest of the addon continues to function correctly.

This fixes an issue in the deeprun tram and other zones with no
coordinates.